gao: A Complete Library of GAO Reports and Metadata

A complete library of reports and metadata from the U.S. Government Accountability Office (GAO). Ships with a bundled dataset of all GAO report URLs and provides functions for downloading reports as PDF or HTML.

Details

This package is not affiliated with or endorsed by the U.S. Government Accountability Office. All data is obtained from public web pages at https://www.gao.gov.

System requirements

Browsing the bundled dataset, exporting metadata, and extracting text from PDFs work out of the box. Downloading report files from gao.gov additionally requires curl-impersonate, a variant of curl that gao.gov's servers accept; functions that need it print installation instructions if it is missing.

See also

Author

Maintainer: Jack T. Rametta jtrametta@gmail.com (ORCID) [copyright holder]

Authors:

Jack T. Rametta jtrametta@gmail.com (ORCID) [copyright holder]